combined providers with weagmiproviders
