--- components/Providers.tsx
@@ -1,13 +1,0 @@
-"use client";
-import { ReactNode } from "react";
-import { ThemeProvider } from "next-themes";
-
-interface ProvidersProps {
-  children: ReactNode;
-}
-
-const Providers: React.FC<ProvidersProps> = ({ children }) => {
-  return <ThemeProvider attribute="class">{children}</ThemeProvider>;
-};
-
-export default Providers;
